requirements.txt
pyyaml
langchain>=0.0.155
fastapi
uvicorn
py3langid
iso-639
openai
opencv-python
gradio
httpx[socks]
icecream
astunparse
scikit-learn
scitools-pyke
nltk
python-constraint
func_timeout
z3-solver
ply
tqdm
